Fixing deploy cp with empty repo file Fixing path from incoming repo

+++ operations/_scripts/deploy/deploy.sh
@@ -29,7 +29,7 @@ export LB_LOGS_BUCKET="$(/bin/bash $GITHUB_ACTION_PATH/operations/_scripts/gener
 mkdir -p "${GITHUB_ACTION_PATH}/operations/deployment/env-files"
 echo "$ENV_GHV" > "${GITHUB_ACTION_PATH}/operations/deployment/env-files/ghv.env"
 echo "$ENV_GHS" > "${GITHUB_ACTION_PATH}/operations/deployment/env-files/ghs.env"
-if [ -s "$GITHUB_WORKSPACE/$ENV_REPO" ]; then
+if [ -s "$GITHUB_WORKSPACE/$ENV_REPO" ] && [ -n $ENV_REPO ]; then
   cp "$GITHUB_WORKSPACE/$ENV_REPO" "${GITHUB_ACTION_PATH}/operations/deployment/env-files/repo.env"
 fi
